[FIRRTL] Move FullReset runner to FullReset pass
